Split styles from pages
