+++ ui/layouts/App/styles.js
@@ -1,0 +1,32 @@
+import styled from 'styled-components';
+
+const App = styled.div`
+  visibility: ${props => (props.ready && !props.loading ? 'visible' : 'hidden')};
+
+  > .container {
+    margin-bottom: 80px;
+    padding-bottom: 20px;
+  }
+
+  .verify-email {
+    margin-bottom: 0;
+    padding: 0;
+    border-top: none;
+    border-bottom: 1px solid #e7e7e7;
+    background: #fff;
+    color: var(--gray-dark);
+    border-radius: 0;
+
+    p {
+      padding: 19px;
+    }
+
+    .btn {
+      padding: 0;
+    }
+  }
+`;
+
+export default {
+  App,
+};
