+++ ui/pages/Documents/styles.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+const Documents = styled.div`
+  table tbody tr td {
+    vertical-align: middle;
+  }
+`;
+
+export default {
+  Documents,
+};
