+++ ui/pages/ExamplePage/content.js
@@ -1,0 +1,4 @@
+export default `
+### This is my Markdown page
+I can type **any** Markdown I want into this file and it will ultimately be parsed into HTML by the &lt;Page /&gt; component in the source of this page.
+`;
