+++ ui/pages/ExamplePage/index.js
@@ -1,10 +1,6 @@
 import React from 'react';
 import Page from '../Page';
-
-const content = `
-### This is my Markdown page
-I can type **any** Markdown I want into this file and it will ultimately be parsed into HTML by the &lt;Page /&gt; component in the source of this page.
-`;
+import content from './content';
 
 const ExamplePage = () => (
   <div className="ExamplePage">
