+++ ui/pages/Index/index.js
@@ -1,79 +1,9 @@
 import React from 'react';
 import { Button } from 'react-bootstrap';
-import styled from 'styled-components';
-import { lighten, darken } from 'polished';
-
-const StyledIndex = styled.div`
-  padding: 20px;
-  background: var(--cb-blue);
-  text-align: center;
-  border-radius: 3px;
-  color: #fff;
-
-  img {
-    width: 100px;
-    height: auto;
-  }
-
-  h1 {
-    font-size: 28px;
-  }
-
-  p {
-    font-size: 18px;
-    color: ${lighten(0.25, '#4285F4')};
-  }
-
-  > div {
-    display: inline-block;
-    margin: 10px 0 0;
-
-    .btn:first-child {
-      margin-right: 10px;
-    }
-
-    .btn {
-      border: none;
-    }
-  }
-
-  footer {
-    margin: 20px -20px -20px;
-    border-top: 1px solid ${darken(0.1, '#4285F4')};
-    padding: 20px;
-
-    p {
-      font-size: 14px;
-      line-height: 22px;
-      color: ${lighten(0.35, '#4285F4')};
-      margin: 0;
-    }
-
-    p a {
-      color: ${lighten(0.35, '#4285F4')};
-      text-decoration: underline;
-    }
-  }
-
-  @media screen and (min-width: 768px) {
-    padding: 30px;
-
-    footer {
-      margin: 30px -30px -30px;
-    }
-  }
-
-  @media screen and (min-width: 992px) {
-    padding: 40px;
-
-    footer {
-      margin: 40px -40px -40px;
-    }
-  }
-`;
+import Styles from './styles';
 
 const Index = () => (
-  <StyledIndex>
+  <Styles.Index>
     <img
       src="https://s3-us-west-2.amazonaws.com/cleverbeagle-assets/graphics/email-icon.png"
       alt="Clever Beagle"
@@ -87,7 +17,7 @@ const Index = () => (
     <footer>
       <p>Want to ensure that your product sees the light of day? <a href="https://cleverbeagle.com?utm_source=pup&utm_medium=app&utm_campaign=oss">Work with Clever Beagle</a>.</p>
     </footer>
-  </StyledIndex>
+  </Styles.Index>
 );
 
 export default Index;
