+++ ui/pages/Profile/styles.js
@@ -1,0 +1,58 @@
+import styled from 'styled-components';
+
+const Profile = styled.div`
+  .nav.nav-tabs {
+    margin-bottom: 20px;
+  }
+
+  .LoggedInWith {
+    padding: 20px;
+    border-radius: 3px;
+    color: #fff;
+    border: 1px solid var(--gray-lighter);
+    text-align: center;
+
+    img {
+      width: 100px;
+    }
+
+    &.github img {
+      width: 125px;
+    }
+
+    p {
+      margin: 20px 0 0 0;
+      color: var(--gray-light);
+    }
+
+    .btn {
+      margin-top: 20px;
+
+      &.btn-facebook {
+        background: var(--facebook);
+        border-color: var(--facebook);
+        color: #fff;
+      }
+
+      &.btn-google {
+        background: var(--google);
+        border-color: var(--google);
+        color: #fff;
+      }
+
+      &.btn-github {
+        background: var(--github);
+        border-color: var(--github);
+        color: #fff;
+      }
+    }
+  }
+
+  .btn-export {
+    padding: 0;
+  }
+`;
+
+export default {
+  Profile,
+};
